+++ android/app/src/main/kotlin/app/sonicsound/IBroadcastObserver.java
@@ -1,5 +1,0 @@
-package app.sonicsound;
-
-public interface IBroadcastObserver {
-    void update(String action, String value);
-}