+++ android/app/src/main/kotlin/app/sonicsound/IBroadcastObserver.kt
@@ -1,0 +1,5 @@
+package app.sonicsound
+
+fun interface IBroadcastObserver {
+    fun update(action: String?, value: String?)
+}